Fix a couple of issues spotted by alexcrichton.
See comments https://github.com/rust-lang/libc/pull/194.

Note on struct utsname: Neither using a constant, nor a conditional
macro seems to work, so I just created an ugly utsname with conditions
on each field. This should also fix the CI fallout. #[cfg] on macros
doesn't work?

As DragonFly only supports one platform, I merged  x86_64.rs into
dragonfly/mod.rs.
